+++ src/TokenGenerator/types.ts
@@ -1,3 +1,4 @@
+/* eslint-disable camelcase */
 export interface Token {
   access_token: string;
   token_type: string;